fix(app): fix react-doctor warnings
