+++ src/components/Chat/components/ChatMessage/renderers/UserChatBody.tsx
@@ -12,7 +12,8 @@ import { ChatNoticeMetaRow } from './ChatNoticeMetaRow';
 import { RichChatMessageUsername } from '../RichChatMessageUsername';
 import { ChatMessageBadges } from './ChatMessageBadges';
 import { ChatMessageBody } from './ChatMessageBody';
-import { canRenderMessageInline, InlineMessageLine } from './InlineMessageLine';
+import { canRenderMessageInline } from '@app/components/Chat/util/canRenderMessageInline';
+import { InlineMessageLine } from './InlineMessageLine';
 import { ReplyingToHeader } from './ReplyingToHeader';
 import type { BadgePressData } from '../RichChatMessage.types';
 import type { UseChatMessagePartRendererArgs } from './useChatMessagePartRenderer';
